Final iterntion added profile tests for all functions 2.added main.h file 3.corrected ftoa function

+++ Project-2/Includes/circbuff.h
@@ -25,7 +25,7 @@ uint8_t IfFull(CircBuff_t *cb);
 uint8_t IfEmpty(CircBuff_t *cb);
 state add_item(CircBuff_t *cb, uint8_t item);
 state remove_item(CircBuff_t *cb, volatile uint8_t *dest);
-void init_buff(CircBuff_t *buff, uint8_t length);
+void init_buff(CircBuff_t *buff, uint32_t length);
 void empty_buff(CircBuff_t *buff);
 void fill_buff(CircBuff_t *buff);
